remove sdx slice name in stitching request
Let SDX manager decide which sdx slice to use. But still keep sdx node optional in stitching request, so that we can still run a sdx slice with multiple sets of routers on the same site and stitch different customer slices to different edge routers in the same slice.

Update readme for the updated code from cnert-2018.