Reactive connection setup triggered by first packet.
Use plexus-v2

When the edge SDX receives prefix notification, it installs monitoring flows
in routers for packets destined for the prefix and from the prefix.
The first packet will be forwarded to sdn controller

Sdn controller notifies SDX controller the event. Sdn controller
find the matching prefix pair for the packet and verify the connection.
If the connection is allowed, install flows.

Previous method by manually requesting for connection still works.